==> FULLSTACK/Notebook_autre_dep/3_meteo_cleaner_autre_dep.ipynb <==

import pandas as pd
import requests
import boto3

# import fichier meteo Ajaccio
path = 13
df = pd.read_csv(f'output/{path}_output.csv', sep=',')
pd.set_option('display.max_columns', None)


df.head(10)

df.info()

# traitement de la date

# Convertir la colonne DATE en datetime si ce n'est pas déjà fait
df['DATE'] = pd.to_datetime(df['DATE'], format='%Y%m%d')
# df['année'] = df['DATE'].dt.year
# df['mois'] = df['DATE'].dt.month
# df['jour'] = df['DATE'].dt.day



# Traitements de colonnes innutiles
drop_cols = list(set([
	'PMERM','PMERMIN','QPMERMIN','FF2M','QFF2M','FXI2','QFXI2','DXI2','QDXI2','HXI2','QHXI2','DXI3S','QDXI3S','DHUMEC','QDHUMEC','INST','QINST','GLOT','QGLOT','DIFT','QDIFT','DIRT','QDIRT','SIGMA','QSIGMA','INFRART','QINFRART','UV_INDICEX','QUV_INDICEX','NB300','QNB300','BA300','QBA300','NEIG','QNEIG','BROU','QBROU','GRESIL','GRELE','QGRELE','ROSEE','QROSEE','VERGLAS','QVERGLAS','SOLNEIGE','QSOLNEIGE','GELEE','QGELEE','FUMEE','QFUMEE','UV','QUV','TMERMAX','QTMERMAX','TMERMIN','QTMERMIN','HNEIGEF','QHNEIGEF','NEIGETOTX','QNEIGETOTX','NEIGETOT06','QNEIGETOT06','QRR','QDRR','QTN','QHTN','QTX','QHTX','QTM','QTMNX','QTNSOL','QTN50','DG','QDG','QTAMPLI','QTNTXM','QPMERM','QFFM','QFXI','QDXI','QHXI','QFXY','QDXY','QHXY','QFXI3S','QHXI3S','QUN','QHUN','QUX','QHUX','QDHUMI40','QDHUMI80','QTSVM','QUM','QORAG','QGRESIL','QBRUME','ECLAIR','QECLAIR','QETPMON','QETPGRILLE'
]))
df = df.drop(columns=drop_cols, axis=1)



# convertion des colonnes object en float64
for column in df.columns:
    if df[column].dtype == 'object':
        # Replace comma with dot for float conversion
        df[column] = df[column].str.replace(',', '.', regex=False)
        df[column] = df[column].astype('Float64')



A = 5
# Convertir la colonne post en code postal à 5 chiffres
df['Code INSEE'] = df['POSTE'].astype(str).str[:A].astype(str)

df.head()

df2 = pd.read_csv('correspondance-code-insee-code-postal.csv', sep=';')
df2.head()


df2 =df2.drop(columns=['Département','Région','Statut','Altitude Moyenne','Superficie','Population','geo_shape','ID Geofla','Code Commune','Code Canton','Code Arrondissement','Code Département','Code Région'])

df_merge = pd.merge(df, df2, on='Code INSEE', how='left')
df_merge.head(10)


df_merge.info()

df_merge['Commune'] = df_merge['Commune'].str.title()
df_merge = df_merge.rename(columns={'Commune' : 'ville'})


df_merge.head()

df_merge.to_csv(f'output/{path}_meteo_cleaned.csv', index=False, sep=';')

# print("Bucket utilisé :", bucket)
# s3 = boto3.client(
#     "s3",
#     aws_access_key_id=aws_access_key_id,
#     aws_secret_access_key=aws_secret_access_key,
#     region_name=region
# )
# s3.upload_file("83/83_meteo_cleaned.csv", bucket, "83_meteo_cleaned.csv")
