==> FULLSTACK/Notebook_autre_dep/4_meteo_features.ipynb <==
import pandas as pd
import numpy as np
import boto3

path = 34
df = pd.read_csv(f'output/{path}_meteo_cleaned.csv', sep=';')
pd.set_option('display.max_columns', None)
df.head()

# fonction de moyenne lissante avec np.convolve
def moving_average(x, w):
    # Remplir le tableau d'entrée avec 'w//2' éléments de chaque côté en utilisant les valeurs de bord
    padded_x = np.pad(x, (w//2, w//2), mode='edge')
    # Effectuer la convolution avec le mode 'valid'
    return np.convolve(padded_x, np.ones(w), 'valid') / w



# ajout de colonne sur les précispitation moyenne par an et mois
df['moyenne precipitations année'] = moving_average(df['RR'], 365).round(2)
df['moyenne precipitations mois'] = moving_average(df['RR'], 31).round(2)

# moyenne ecapotranspiration par mois et année
df['moyenne evapotranspiration année'] = moving_average(df['ETPMON'], 365).round(2)
df['moyenne evapotranspiration mois'] = moving_average(df['ETPMON'], 31).round(2)

# moyenne vitesse de vent par mois et année
df['moyenne vitesse vent année'] = moving_average(df['FFM'], 365).round(2)
df['moyenne vitesse vent mois'] = moving_average(df['FFM'], 31).round(2)

# moyenne température par mois et année
df['moyenne temperature année'] = moving_average(df['TN'], 365).round(2)
df['moyenne temperature mois'] = moving_average(df['TN'], 31).round(2)

df.to_csv(f'output/{path}_meteo_cleaned_features.csv', sep=';', index=False, encoding='utf-8')

# print("Bucket utilisé :", bucket)
# s3 = boto3.client(
#     "s3",
#     aws_access_key_id=aws_access_key_id,
#     aws_secret_access_key=aws_secret_access_key,
#     region_name=region
# )
# s3.upload_file("corse_meteo_cleaned_insee_features.csv", bucket, "corse_meteo_cleaned_insee_features.csv")
