==> FULLSTACK/Notebook_autre_dep/5_Fusion_feu-meteo_autre_dep.ipynb <==
import pandas as pd
import requests
import time
import os
import matplotlib.pyplot as plt
import plotly.express as px
import plotly.graph_objects as go
import numpy as np
import seaborn as sns
from geopy.distance import geodesic
from sklearn.cluster import DBSCAN
from sklearn.pipeline import Pipeline
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import OneHotEncoder
from tqdm import tqdm
from geopy.distance import geodesic

#### Visu de l'historique des feux

path = 13
feux = pd.read_csv('https://projet-incendie.s3.eu-west-3.amazonaws.com/historique_incendies_avec_coordonnees.csv', sep=';', encoding='utf-8')
pd.set_option('display.max_columns', None)

feux.head(10)

mask = feux['Département'] == f'{path}'
feux_ville = feux[mask]
feux_ville

#### Rajout de la météo


df_meteo=pd.read_csv(f'output/{path}_meteo_cleaned_features.csv', sep=';', encoding='utf-8')
df_meteo.head()

df_meteo.rename({'DATE': 'Date'}, axis=1, inplace=True)
df_meteo['Date'] = pd.to_datetime(df_meteo['Date']).dt.normalize()

# df_meteo = df_meteo.rename(columns={
#     "année": "year",
#     "mois": "month",
#     "jour": "day"
# })

# # Étape 2 : Créer une colonne datetime à partir de ces 3 colonnes
# df_meteo["date"] = pd.to_datetime(df_meteo[["year", "month", "day"]])

df_meteo.head()

df_meteo['Code INSEE'] = df_meteo['Code INSEE'].astype(str)

feux_ville['Date'] = pd.to_datetime(feux_ville['Date']).dt.normalize()




df_meteo['Code Postal'] = df_meteo['Code Postal'].astype(str)

### Merge Feux et météo

feux_ville = feux_ville.rename(columns={'Nom de la commune': 'ville'})
df= pd.merge(df_meteo, feux_ville, on=['Date', 'ville'], how='outer')

df.head()

df['Feux'].value_counts()

### Traitement des doublons

# df_clean = df.groupby(['ville', 'Date'], as_index=False).agg(lambda x: x.dropna().iloc[0] if not x.dropna().empty else None)

df['Feux'] = df['Feux'].fillna(0)
df['Feux'].value_counts()

# # # Supprime aussi les tabulations, retours chariots, etc.
# # df.columns = df.columns.str.replace(r"\s+", " ", regex=True).str.strip()
# # Filtrer les communes sans données météo mais AVEC des coordonnées valides
# df_manquante = df_manquante[df_manquante["latitude"].notna() & df_manquante["longitude"].notna()]

# # Pareil côté communes valides
# df_valide = df_valide[df_valide["latitude"].notna() & df_valide["longitude"].notna()]

# def trouver_commune_la_plus_proche(row, df_sources):
#     coord_source = (row["latitude"], row["longitude"])
#     distances = df_sources.apply(
#         lambda r: geodesic(coord_source, (r["latitude"], r["longitude"])).km, axis=1
#     )
#     return df_sources.loc[distances.idxmin()]

# colonnes_meteo = ["RR", "TX", "TM", "ETPMON", "UM",'RR','DRR','TN','HTN','TX','HTX','TM','TMNX','TNSOL','TN50','TAMPLI','TNTXM','FFM','FXI','DXI','HXI','FXY','DXY','HXY','FXI3S','HXI3S','UN','HUN','UX','HUX','DHUMI40','DHUMI80','TSVM','UM','ORAG','BRUME','ETPMON','ETPGRILLE','Code INSEE_x','Code Postal','Commune_x','geo_point_2d_x','moyenne precipitations année','moyenne precipitations mois','moyenne evapotranspiration année','moyenne evapotranspiration mois','moyenne vitesse vent année','moyenne vitesse vent mois','moyenne temperature année','moyenne temperature mois']  # ou toute autre liste pertinente


# # Boucle de remplissage
# for i, row in df_manquante.iterrows():
#     plus_proche = trouver_commune_la_plus_proche(row, df_valide)
#     for col in colonnes_meteo:
#         if pd.isna(row[col]):
#             df_manquante.at[i, col] = plus_proche[col]

# df_complet = pd.concat([df_valide, df_manquante], ignore_index=True)

df.to_csv(f"output/{path}_Dataset_modele.csv", index=False, sep=';')
