==> FULLSTACK/Notebook_autre_dep/6-1_requete_gps_ville.ipynb <==
import pandas as pd
import requests
import time
import tqdm
import io
import json

path = 34
df_nombre = pd.read_csv(f'output/{path}_nombre_unique')
df_nombre = df_nombre.drop_duplicates(subset='ville')
display(df_nombre.describe(include='all'))

ajaccio = []
for i in df_nombre['ville']:
    reponse = requests.get(f"https://api-adresse.data.gouv.fr/search/?q={i}")
    reponse = reponse.json()
    ajaccio.append(reponse)
    time.sleep(60 / 300)


# Flatten the 'features' from each response in ajaccio and create a DataFrame
features_list = []
for response in ajaccio:
	if 'features' in response:
		features_list.extend(response['features'])

df_features = pd.json_normalize(features_list)
df_features.head()

df_features[['longitude', 'latitude']] = pd.DataFrame(df_features['geometry.coordinates'].tolist(), index=df_features.index)

df_features.info()

df_features = df_features.drop(df_features.columns[[0,1,2,3,4,6,7,9,10,11,12,13,14,15,16,18,19,20,21]], axis=1)

df_features.head()

df_features['properties.municipality'].dropna()
df_features['properties.municipality'] = df_features['properties.municipality'].str.title()
df_features['properties.name'] = df_features['properties.name'].str.title()
df_features.head()

df_filtre = df_features[df_features['properties.id'].str.startswith((f'{path}'))&
    (df_features['properties.id'].str.len() <= 5)]

display(df_filtre.describe(include='all'))

df_filtre.to_csv(f"output/{path}_new_gps.csv")
